i#2364: adding test case to retaddr instead of creating a new one
