+++ suite/tests/client-interface/retaddr.dll.c
@@ -32,15 +32,36 @@
 
 #include "dr_api.h"
 
+uint num_ret = 0;
+
+static void mbr_instru(app_pc instr_addr, app_pc target_addr) {
+    num_ret++;
+}
+
 static dr_emit_flags_t
 bb_event(void *drcontext, void *tag, instrlist_t *bb, bool for_trace, bool translating)
 {
     instr_t *instr;
     for (instr = instrlist_first(bb); instr != NULL; instr = instr_get_next(instr)) {
-        if (instr_is_return(instr))
+        if (instr_is_return(instr)) {
             dr_clobber_retaddr_after_read(drcontext, bb, instr, 0);
+            /* i#2364 : mbr should be tested against different kinds of return. */
+            dr_insert_mbr_instrumentation(drcontext, bb, instr,
+                                          (app_pc)mbr_instru, SPILL_SLOT_1);
+        }
     }
     return DR_EMIT_DEFAULT;
+}
+
+void exit_event(void)
+{
+    /* We assume returns are executed at least twice. */
+    if (num_ret > 1) {
+        dr_fprintf(STDERR, "instrumentation for return ok\n");
+    }
+    else {
+        dr_fprintf(STDERR, "FAIL no instrumentated returns\n");
+    }
 }
 
 DR_EXPORT
@@ -48,4 +69,5 @@ void dr_init(client_id_t id)
 {
     dr_fprintf(STDERR, "thank you for testing the client interface\n");
     dr_register_bb_event(bb_event);
+    dr_register_exit_event(exit_event);
 }
