+++ suite/tests/client-interface/retaddr.expect
@@ -4,3 +4,4 @@ retaddr 0x0
 retaddr 0x0
 retaddr 0x0
 All done
+instrumentation for return ok
