--- suite/tests/client-interface/retinstru.expect
@@ -1,7 +1,0 @@
-thank you for testing the client interface
-in callee
-return
-return
-return
-All done
-instrumentation for return ok
